Extrais un composant <SommaireSynchroAvecScroll />
… qui sera réutilisé sur la page "Me préparer à NIS 2"
